requirements.txt
greenlet==3.1.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
typing_extensions==4.12.2
python-dotenv~=1.0.1